backend/requirements.txt
protobuf 
dialogflow 
authlib 
djangorestframework 
google 
google-cloud-vision 
google-api-core 
django 
googletrans 
django-rest-framework-mongoengine 
mongoengine 
pandas 
owlready2 
django-requestlogs 
django-ipware 
opencv-python 
scipy 
ordered-set 
matplotlib 
webcolors 
pillow 
boto3 
tensorflow 
blinker 
num2words 